Finished Player Appearance

+++ src/main/java/com/michaelgatesdev/OldLeaf/game/player/appearance/Characteristic.java
@@ -20,68 +20,69 @@
 
 package com.michaelgatesdev.OldLeaf.game.player.appearance;
 
+import com.michaelgatesdev.OldLeaf.game.player.Gender;
 import javafx.scene.image.Image;
 
 public abstract class Characteristic
 {
     // ============================================================================================================================================ \\
-
+    
     private Image  thumb;
     private Gender gender;
-    private Short  value;
-
+    private Byte   value;
+    
     // ============================================================================================================================================ \\
-
-
-    public Characteristic(Short value, Gender gender, Image thumb)
+    
+    
+    public Characteristic(Byte value, Gender gender, Image thumb)
     {
         this.thumb = thumb;
         this.gender = gender;
         this.value = value;
     }
-
+    
     // ============================================================================================================================================ \\
-
-
+    
+    
     public void setThumb(Image thumb)
     {
         this.thumb = thumb;
     }
-
-
+    
+    
     public void setGender(Gender gender)
     {
         this.gender = gender;
     }
-
-
-    public void setValue(Short value)
+    
+    
+    public void setValue(Byte value)
     {
         this.value = value;
     }
-
-
+    
+    
     // ============================================================================================================================================ \\
-
-
+    
+    
     public Image getThumb()
     {
         return thumb;
     }
-
-
+    
+    
     public Gender getGender()
     {
         return gender;
     }
-
-
-    public Short getValue()
+    
+    
+    public Byte getValue()
     {
         return value;
     }
-
-
+    
+    
     // ============================================================================================================================================ \\
 }
