+++ src/main/java/com/michaelgatesdev/OldLeaf/game/player/appearance/EyeColor.java
@@ -20,11 +20,12 @@
 
 package com.michaelgatesdev.OldLeaf.game.player.appearance;
 
+import com.michaelgatesdev.OldLeaf.game.player.Gender;
 import javafx.scene.image.Image;
 
 public final class EyeColor extends Characteristic
 {
-    public EyeColor(Short value, Gender gender, Image thumb)
+    public EyeColor(Byte value, Gender gender, Image thumb)
     {
         super(value, gender, thumb);
     }
