+++ src/main/java/com/michaelgatesdev/OldLeaf/game/player/appearance/FaceStyle.java
@@ -20,11 +20,12 @@
 
 package com.michaelgatesdev.OldLeaf.game.player.appearance;
 
+import com.michaelgatesdev.OldLeaf.game.player.Gender;
 import javafx.scene.image.Image;
 
 public final class FaceStyle extends Characteristic
 {
-    public FaceStyle(Short value, Gender gender, Image thumb)
+    public FaceStyle(Byte value, Gender gender, Image thumb)
     {
         super(value, gender, thumb);
     }
