+++ src/main/java/com/michaelgatesdev/OldLeaf/game/player/appearance/HairColor.java
@@ -20,11 +20,12 @@
 
 package com.michaelgatesdev.OldLeaf.game.player.appearance;
 
+import com.michaelgatesdev.OldLeaf.game.player.Gender;
 import javafx.scene.image.Image;
 
 public final class HairColor extends Characteristic
 {
-    public HairColor(Short value, Gender gender, Image thumb)
+    public HairColor(Byte value, Gender gender, Image thumb)
     {
         super(value, gender, thumb);
     }
