+++ src/main/java/com/michaelgatesdev/OldLeaf/game/player/appearance/HairStyle.java
@@ -20,11 +20,12 @@
 
 package com.michaelgatesdev.OldLeaf.game.player.appearance;
 
+import com.michaelgatesdev.OldLeaf.game.player.Gender;
 import javafx.scene.image.Image;
 
 public final class HairStyle extends Characteristic
 {
-    public HairStyle(Short value, Gender gender, Image thumb)
+    public HairStyle(Byte value, Gender gender, Image thumb)
     {
         super(value, gender, thumb);
     }
